feat(cycle): Добавлена блокировка запуска без контракта

+++ src/Host/Controllers/Colonies/ColonyDetails.cs
@@ -5,7 +5,7 @@
         long UserId,
         string Name,
         int SolarsIncome,
-        int GavernorType,
+        double GavernorType,
         int Population,
         int ZonesOccupied)
         : ColonySummary(
